requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
ddtrace==4.3.1
kafka-python==2.3.0
structlog==25.5.0
ecs-logging==2.3.0
pydantic==2.12.5
pydantic-settings==2.12.0
tomli==2.4.0
